Add ResultsCount component
This is a copy of the results count portion of the results header
component. Separating this out allows for basic yet priorly
exhausting combinations. In the past, even wanting to just flip
the order of the results count and applied filters, or insert
simple content in between them, would require a full override
of the results header template.

TEST=manual

This is used in the collapsible filters page
See http://engblog.yext.com/collapsible-filters-prototype/people.html