Map: add shouldFreezeMap config function
This commit adds the shouldFreezeMap config function, which
allows the user to specify a criteria when the map should not update.

TEST=manual

this is used at http://engblog.yext.com/collapsible-filters-prototype/people.html
to not update the map on mobile phones (since it is hidden in those cases)
Otherwise, for low-end phones, we noticed there was a chance to crash the
page if a user made too many searches too quickly. It also just helps
performance a lot