Add the Stylelint CSS linter and some initial rules. (#1076)
This PR adds the Stylelint linter to the SDK. This linter can parse CSS and
CSS-like (SCSS) syntax. It has built-in rules and an ecosystem of plugins
that will allow us to catch potential styling errors and enforce best CSS
practices. As part of this PR, I added a few of their recommended rules. We
will continue to add rules as time goes on. The SDK's SCSS did not adhere
to these rules, so I fixed the files that violated them.

The 'no-duplicate-selectors' seems to have an issue where it considers the
selectors 'a' and 'a + a' to be the same. That's why I disabled it for the
_Card.scss file. Even with this bug, I think the rule is valuable to add. The
'immeadiately after' selector is not one we use often. We can easily disable
this rule when we need to.

Stylelint now runs as part of 'npm run test'.

J=NONE
TEST=auto

Ensured the Stylelint checks passed after updating the offending SCSS files.