Ensure all styling in the SDK is self contained. (#1071)
The SDK's styling should affect only the components on the page and nothing else. Unfortunately, we had some global CSS resets that were applied to an entire page, not just the Answers components. This PR corrects that. In order to scope these resets, we added a CSS class name that will be applied to all component containers. The class name chosen was 'yxt-Answers-component'. I also correctly scoped some styling for the Pagination component.

In the future, we should add a CSS linter to the SDK. We should add a linting rule that will detect any styling not scoped to the components.

J=SLAP-571
TEST=manual

Ensured that the CSS resets were still applied to all components. Verified that I could override a reset within a component's styling.