filteroptions: Fix back navigation behavior by applying previous state
Consider the following answers navigation situation:

In State 1
Search "all"
In State 2
Click on a static filter (e.g. distance "5 miles")
In State 3
Navigate backwards in browser history
In State 4

You would expect state 4 == state 2. However, this is currently not the case.
The results from state 3 still exist in state 4 instead of the results from state 2.
The URL in state 4 reflects the same URL as state 2, however the results are
different. This is because we are not applying the filter changes from the backwards
navigation, even though it is being changed in globalStorage.

This PR aims to update the FilterOptions component when it is updated on a
backwards navigation (ie in globalStorage). We do not fix other static filters like
RangeFilter or DateRangeFilter or SortOptions as that will be part of a bigger rework
aiming to get all Facets/Static Filters manifesting on load/backwards navigation.

J=SLAP-624
TEST=manual

Test on a site using a local SDK dist directly

Test a FilterBox component with RADIUS_FILTER
Test a FilterBox component with STATIC_FILTER
Test a Facet component

Click on a static filter option. Navigate backwards in history. Make sure
the results AND the url reflect correctly in the back nav (that is, the filter
value changes in the URL and the results show the previous state).

Try going from FilterBox.filter0=[[something]] to FilterBox.filter0=[[nothing]]
Try going from FilterBox.filter0=[[something]] to FilterBox.filter0=[[somethingelse]]