Add results count component (#1079)
Add ResultsCount component

This is a factoring out of the results count portion of the results header
component. Separating this out allows for basic yet priorly
exhausting combinations. In the past, even wanting to just flip
the order of the results count and applied filters, or insert
simple content in between them, would require a full override
of the results header template.

We can't update the results header component to use this yet,
due to more fundamental issues the sdk has with nested components.

TEST=manual, auto

added jest tests

This is used in the collapsible filters page
See http://engblog.yext.com/collapsible-filters-prototype/people.html