Add Applied Filters Component (#1078)
Add Applied Filters Component

This is a factoring out of the applied filters portion of the results header
component. Separating this out allows for basic yet priorly
exhausting combinations. In the past, even wanting to just flip
the order of the results count and applied filters, or insert
simple content in between them, would require a full override
of the results header template.

We can't use this component inside the results header component,
because the SDK is bad at nesting components. Specifically, when
a component is removed (which happens every time a child component
is rerendered) it will remove its moduleId event listener, which has the
side effect of removing all other listeners on that same moduleId.

TEST=manual

This is used in the collapsible filters page
See http://engblog.yext.com/collapsible-filters-prototype/people.html

also added it to a standalone page, and saw that it was the removable filters worked, and it updated at the end of every search