Add a resultsCountByVertical entry to the onUniversalSearch input. (#1082)
This PR adds a new entry to the Object passed to onUniversalSearch. This entry,
called resultsCountByVertical, is an Object with values of the form:

{ totalResultsCount: 150, displayedResultsCount: 10}

for each returned vertical. The totalResults count indicates how many results
were found for that vertical. The displayedResultsCount indicates how many
of those results are actually displayed. This additional info was requested
by Mercy. But, it seemed generally useful, so we're adding it to the SDK
officially.

J=SLAP-648
TEST=manual

On a local test site, logged the param passed to onUniversalSearch. I verified
that the resultsCountByVertical was correct by comparing it against the
LiveAPI response.