VerticalResultsCount: use VERTICAL_RESULTS storage key instead of RESULTS_HEADER
VerticalResultsCount cannot use StorageKeys.RESULTS_HEADER while a
VerticalResults component (that does not have hideResultsHeader: true)
is on the page, because of the bug that when an event listener is unregistered
(as happens with child components like the ResultsHeader inside VerticalResults)
it will unregister all other listeners for that storage key.

TEST=manual

Tested that I can have both a VerticalResults (with its own results header)
and a VerticalResultsCount on the same page, and both will update.