autocomplete: add wcag attributes to hbs
Previously in #1034 for 1.4. Have pivoted to have these changes in
develop (minor version) release instead.

For the autocomplete list, we add a variety of labels and roles to
better support screen reader identification.

We add a results count to the autocomplete with an aria-live="assertive"
so this attribute is called as updates are given.

J=SLAP-537
TEST=manual

Test on a local Jambo HH Theme repository with one searchbar.
Test that autocomplete items have unique IDs
Test that autocomplete total number of results is present in the DOM and
is accurate as you change the autocomplete results.