Update translation helper interfaces (#1084)
Update the translation helpers to no longer use stringified JSON

Update the TranslationProcessor to expect an object rather than stringified JSON for plural forms. Update the HBS helper to find pluralForm keys and and pass them to the TranslationProcessor in an object as the first parameter. Update fallback behavior surrounding TranslationProcessor by getting the locale if the option is specified and otherwise using the locale from the init. Get the language as first two characters of the locale. If it is supported, use that when selecting plural forms. Otherwise use 'en'.

J=SLAP-638
TEST=auto, manual

Update unit tests and confirm they pass. Run an internationalized jambo build and observe the translations working.