Update translator interface (#1096)
Update the interface of the translate helpers, and update the SDK build step to match it

Update the TranslationProcessor to expect an object rather than stringified JSON for plural forms. Update the HBS helper to find pluralForm keys and and pass them to the TranslationProcessor in an object as the first parameter. Update fallback behavior surrounding TranslationProcessor by getting the locale if the option is specified and otherwise using the locale from the init. Get the language as first two characters of the locale. If it is supported, use that when selecting plural forms. Otherwise use 'en'.

In the SDK, update the build step to output runtime processTranslation calls that match this new interface.

J=SLAP-638
TEST=auto, manual

Update unit tests and confirm they pass. Run a build and observe the properly formatted calls to the runtime translation helper methods. Run an internationalized jambo build and observe the runtime translations working.