Use translated templates in jest tests instead of raw templates
I noticed that my unit tests on a separate pr were breaking, and for
some reason were using the raw component template (found by
adding an implementation for the {{translate}} helper) rather than
using the translated template bundle which uses {{processTranslation}}

This commit updates our jest tests to use the translated templates
inside dist/, rather than manually loading in the raw templates.
This simplifies our unit tests, but also makes it a requirement
to rebundle the templates to run a unit test (which is now
the case with the localized-assets changes anyways).

TEST=auto

re ran unit tests