translationprocessor: ie11 fix for regexp constructor
IE11 does not allow you to use a regex literal as the first
argument to the RegExp constructor. I tried targeting ie11
in bundle.js's legacyBundle babel config, but that didn't
seem to fix the issue.

J=SLAP-709
TEST=manual

Test that I can load a page that with components that use
translations (e.g. VerticalResutlsCount) again