AppliedFilters: listen to new VERTICAL_SEARCH_LOADED key
The AppliedFilters components must load after the Facets
component. This is because the Facets component, in addition
to rendering the facets, is also responsible for registering
each new vertical search's new set of facets.

For some context, each vertical search response is capable of
returning a set of facet filters, which we want displayed both
in the AppliedFilters component, as well as the Facets component.
The Facets component is ALSO responsible for loading these
facets into the SDK's FilterRegistry (it should not be).

Created SLAP-721 for correctly dealing with this. It would
be a pretty large code change, so for now we're adding this
short-term fix.

TEST=manual

test that I can search "Consulting", which will return a
search response with the "Consulting" facet selected,
and the AppliedFilters component will show a removable
"Consulting" filter, and that this does not occur
without this change