AppliedFilters: listen to new FACETS_LOADED key (#1110)
This commit adds a FACETS_LOADED global storage key, which
sends an 'update' event when facets are rendered from the response,
and updates the AppliedFilters component to listen to this key
in addition to VERTICAL_RESULTS.

Prior to this change, AppliedFilters had to have its addComponent AFTER the Facets
component. This is because the AppliedFilters
component and Facets component were both listening to the same
StorageKey (DYNAMIC_FILTERS), so the order in which their
moduleId listeners ran depended on the addComponent order.
The Facets component has to run first, though, because in addition
to rendering the facets, it is also responsible for registering
each new vertical search's new set of facets into FilterNodes stored
in FilterRegistry.

For some context, each vertical search response is capable of
returning a set of facet filters, which we want displayed both
in the AppliedFilters component, as well as the Facets component.
The Facets component is ALSO responsible for loading these
facets into the SDK's FilterRegistry (it should not be).

Created SLAP-721 for correctly dealing with this. It would
be a pretty large code change, so for now we're adding this
short-term fix.

J=SLAP-712
TEST=manual

test that I can search "Consulting", which will return a
search response with the "Consulting" facet selected,
and the AppliedFilters component will show a removable
"Consulting" filter, regardless of whether I add AppliedFilters
before or after Facets inside the onReady()