DefaultTemplatesLoader: Fix async template loading race condition
If somebody was using the built-in template loader, instead of
manually specifying a script tag with the templates, if the master
switch api call finished before the templates were loaded,
components would be addComponent-ed without any templates registered.

This commit uses Promise.all() so that we don't lose any page speed.
Originally it just ran the template loader, then the ponyfill,
then the master switch api, but this caused an additional ~75-100ms
load time increase. Updating the code to use Promise.all() caused
the code to go back to it's old performance, barring a few occasional
spikes (which I'm assuming would have errored out without this change)
This was measured by adding a performance mark at the start of the onReady().

There was some consideration to use Promise.race with the master switch api,
so that if the master switch was disabled then the site would fail immediately
instead of waiting for the other async deps, but it felt weird to optimize for
this and it made the code more complicated.

T=https://yextops.zendesk.com/agent/tickets/356651
TEST=manual

test that if I force the template loader to wait an extra 2 seconds
before creating the templates script tag, the rest of the page will
wait for the templates to load (and check that without this code
change we would see the page fail to load without waiting)

test that if that master switch is disabled, the experience will
still not load

test that if the master switch takes longer than the timeout,
the page will load (tested by changing the timeout to 0ms)

test that I can chain more .then()s after the ANSWERS.init()
promise (was worried that using finally() would prevent this)

test speed as noted above