Verify all translations are present in CircleCI (#1135)
Add a translation_test job to Circle which tests that all the PO files contain all the translations defined in the POT file.

Use the utility "msgcmp" to compare the files.

The original PR for this is #1124 , but I accidentally merged that without squashing, so I am re-doing it.

J=SLAP-724
TEST=manual

Observe CircleCI pass when all translations are present. Push a test commit where a translation is missing and observe it fail.