Add unit tests to the i18n build process (#1129)
Add tests to the code responsible for transpiling the javascript and hbs helpers into processTranslation calls.

Test coverage includes TranslationResolver, TranslateCallParser, TranslateHelperVisitor, Translator, fromMustacheStatementNode(), and generateProcessTranslationJsCall().

This PR also adds tests for the translation extractor and adds pluralWithContext support to the translation resolver. The tests for the translator are from the Jambo translator test suite. I tweaked the ALL_LANGUAGES to make it work with the tests.

J=SLAP-700
TEST=auto

Run unit tests and confirm they pass