Add Percy Integration (#1148)
This commit adds a percy_snapshot github action,
that uploads snapshots of our acceptance test html fixtures
to percy to check for visual changes. A PERCY_TOKEN secret under
the yext/answers repo settings will be necessary for this to work
(tested by just hardcoding the env var). 

To explain why github actions were used instead of adding to our circle config,
our git strategy is more complicated than what Percy can handle without PR integration.
We have multiple "master" percy branches (support, release, develop,
and master), but percy only lets you set one "master" branch. Because of this, in
order to use Percy properly, we have to use it with PR integration turned in.
In order for a percy build to have PR integration, a PR for that commit has to
exist before the percy stuff is run. However, circleci does not have an "on
pull request hook", and builds are run immediately after pushing to the remote.
[apparently other people want this too](https://ideas.circleci.com/cloud-feature-requests/p/trigger-new-build-when-a-pull-request-is-opened)
This creates an unfortunate real life race condition, where if you don't create a PR before
the percy step is run, you will not get PR integration with your percy build. We could
add a hold step to our circleci builds, to give people time to create a PR, or maybe do
something with a webhook, but it seemed simpler to just create a github action for this instead.

J=SLAP-291
TEST=manual

tested that percy builds are run when a pull request is made, for any branch
tested that, for branches that don't match `dev/*`,  percy builds are automatically
run and approved. This is needed so that PRs will have base builds to compare to.
tested that merging in a `dev/*` branch into a non `dev/*` branch (like `develop` or `master`) triggers a 
percy build, which can form the base build for future PRs