Add onSubmit function to SearchBar component (#1132)
Add an onSubmit hook to the SearchBar component. This onSubmit function is necessary for the Overlay and should happen as (or before) the search network call is conducted.

This is different than the onUniversalSearch and onVerticalSearch because those functions are meant to be used for analytics. Those functions happen after the search is conducted and provide result information to a user; they also are expected to return an analytics event.

TEST=manual

Add a simple onSubmit function with a console.log statement. See output printed to console when search is conducted.