SearchBar: Update to show clear button on external query (#1162)
Since this PR (#1130), searches can be triggered from outside of the SearchBar component using a function ANSWERS.search. However, the SearchBar component was previously not adding the "clear button" icon when this happened. Updated the SearchBar component to show the clear button for external queries as well as internal ones.

TEST=manual,unit

Test with the Overlay prototype which uses the ANSWERS.search function. Also tested manually typing in the search input and see clear button appear.