Analytics: Add query source config option (#1163)
Add an analytics query source config option to the ANSWERS.init function; this is passed in the universal and vertical search requests as "source". The analytics "source" key will be used to differentiate analytics data between integration types. Current types accepted by the back end are "OVERLAY" and "STANDARD"

J=SLAP-768
TEST=manual

Specify this option and conduct searches for both vertical and universal. Confirm with the back end analytics teams that they are seeing the source param come through for the request queryIds.