Move non-ie11 acceptance tests into container
Currently we're capped at 2 parallel threads for browserstack automate.
This makes it really annoying when multiple people are trying to debug
acceptance tests.

The only browser we really need browserstack for is ie11.
We could also investigate whether it's possible to get a
windows circleci image.

Also this should improve the speed of our builds

TEST=auto

future me will see tests run properly in build and browserstack