Run firefox and chrome acceptance tests in headless mode
Currently we're capped at 2 parallel threads for browserstack automate.
This makes it really annoying when multiple people are trying to debug
acceptance tests.

If we run the chrome and firefox acceptance tests in headless mode in
the circleci container, we can speed up our builds a lot. Testcafe
only supports firefox and chrome in headless mode.

TEST=auto

see tests run properly in the circleci build