Search: update search hook placement (#1168)
Move custom hook from onQuerySubmit to search. This means that the hook is invoked whenever a search is conducted. Since the hook has not been in any releases, we're free to change its name without any backwards compatibility issues. Changed from onSubmit to onConductSearch.

TEST=manual,unit

Test with Overlay locally. Confirm that this hook is invoked when searches are conducted via typing in the search bar as well as with the ANSWERS.search function.