Sort options focus state (#1165)
Add focus state to the sort options

This PR allows users to tab into the SortOptions component and navigate through the options with the arrow keys (The W3C spec). In order to make this possible, the SortOptions component had to be modified so that it did not re-render when a different sort option was selected.

Note: To get this feature to work on Safari, you must go to "Preferences -> Accessibility" and select "Press Tab to highlight each item on a webpage"

J=SLAP-727
TEST=manual, acceptance

Test that you can tab into the sort options component and navigate the options with the arrow keys on Chrome, Firefox, Safari, and IE11. Test that this change is compatible with the searchOnChange and the showReset options. Test this feature on an Answers site using the Hitchhiker's Theme and Collapsible Filters. Add an acceptance test that confirms that focus works on options.