Create the new CoreAdapter wrapper.
This PR creates the `CoreAdapater`, which will use the new Core library to
perform all network requests. The adapter is essentially the SDK's old
`Core` class, with the addition of an `init` method. This method calls the new
`provideCore` function to get an instance of the library. Note that the Core
library is not added to the SDK's package.json. This is because it has not yet
been published to npm and the Github repo does not contain a dist build.

You can point the SDK at your local version of answer-core by adding a line
like this to the package.json dependencies:

`"@yext/answers-core": "file:../answers-core"`

This will enable you to test your answers-core changes against the SDK. Once
answers-core has been published to npm, we will commit the dependency to the
package.json. We will also be able to get rid of the answers-core specific
things added to bundle.js

I initially considered having answers-umd invoke `provideCore` and then
instantiate `CoreAdapter` after `_loadAsyncDependencies`. This would not have
worked, however. It would mean that `ANSWERS` instance methods, such as
`setSessionsOptIn` could not be called until `onReady`. This would be a
breaking change.

J=SLAP-836
TEST=manual

Pointed the SDK to my local answers-core build. Created a bundle of the SDK
and made sure that `CoreAdapter` was initialized with a Core library instance.