package.json: Add prepublishOnly and postpublish scripts for README (#1183)
* package.json: Add prepublishOnly and postpublish scripts for README

We want to publish to NPM for the first time. There is a problem because
NPM uses the root-level README.md as its source for the package README.
Our current root-level README has langauge that is specific to the
GitHub repository. Thus, we want to provide a shorter version of the
README to NPM that will link to the current GitHub README.

To do this, before any time we do an `npm publish`, we want to switch
out the root-level README with an NPM README, and then postpublish,
switch things back to the way they are currently.

J=SLAP-812
TEST=manual

Tested on a test NPM package and successfully pushed a changed README.md
to the NPM registry and saw it reflected there.

* Executables are runnable

* Use conf not ci (this is what I used in my test)

* Typo in postpublish