SDK changes for NPM public release (#1187)
We want to publish to NPM for the first time. We make
changes to get us ready for the NPM registry (mostly
changes to what is put into the registry and the package
copies.

#1183
#1184

J=SLAP-812
TEST=manual

Tested all PRs on a test NPM package pushed to the
registry under my username (npm init, npm publish).

* package.json: Add prepublishOnly and postpublish scripts for README (#1183)

* package.json: Add prepublishOnly and postpublish scripts for README

We want to publish to NPM for the first time. There is a problem because
NPM uses the root-level README.md as its source for the package README.
Our current root-level README has langauge that is specific to the
GitHub repository. Thus, we want to provide a shorter version of the
README to NPM that will link to the current GitHub README.

To do this, before any time we do an `npm publish`, we want to switch
out the root-level README with an NPM README, and then postpublish,
switch things back to the way they are currently.

J=SLAP-812
TEST=manual

Tested on a test NPM package and successfully pushed a changed README.md
to the NPM registry and saw it reflected there.

* Executables are runnable

* Use conf not ci (this is what I used in my test)

* Typo in postpublish

* package.json: Changes for public npm release (#1184)

* package.json: Changes for public npm release

Change the license to match the LICENSE file.
Change the main file to the dist umd file, as that is the file most
likely to be imported when including this package.

Add the repository, bugs, and homepage fields to point to the current
GitHub repository.

Add a .npmignore for the registry that is exactly like the .gitignore
except it allows the dist folder. We allow the dist on publish so that
we do not need to build on postinstall. The dist files are available as
soon as installation is done.

J=SLAP-812
TEST=manual

Tested on a test NPM package
Saw all changes reflected on the test package in the registry after
pushing.

* Changes from Rose: keywords, author

* Specify node 12 in package.json

* Remove app from npmignore, doesn't exist