Update Core's verticalSearch method to always pass query to onVerticalSearch.
The `verticalSearch` method was previously passing `query.input` to the
`onVerticalSearch` hook. However, this value is undefined when pagination is
used to conduct the search. When `query.input` is undefined, the query in
`globalStorage` should be passed instead.

J=SLAP-828
TEST=manual

Console logged the input to `onVerticalSearch` saw that that `input` was
correct when using the search bar or pagination.