Use extracted core question submission service
This commit removes the previous question submission
service code, and uses the new extracted core one.

The qa submission component does not use the data
returned from the endpoint, and the only data returned
is the uuid of the request and any errors.
Currently there is an issue where the component will
ignore any errors returned by the backend, as long
as no error is thrown.

https://github.com/yext/answers/issues/570

This will be addressed in a later pr, once we decide
how to do errors in the new extracted core.

J=SLAP-838
TEST=manual

tested that I can use the qa submission component
added a console.log to the extracted core for sanity's sake