Location Bias Fix (#1195)
Fix a bug where the location bias component would load improperly

The bug occurred when navigating to a blank vertical, performing a search, and hitting the back button. The location bias component would then have the properties for its state set to null. This caused the component to load since the component was only checking that the accuracy was not undefined. This commit makes sure the state is not null before loading the component.

J=SLAP-574
Test=manual

Verify that the bug described above no longer occurs. Test other searches like 'locations near me' and check that the location bias component still works as expected