Adapt universal search (#1191)
Add adapters that convert the UniversalSearchResponse from answers-core into an SDK `UniversalResults` model

The commented-out lines in the core-adapter represent work that will be completed in future PRs. Future PRs will also add tests for these adapters.

In order to use this, "@yext/answers-core": "file:../answers-core", needs to be added to the package.json and answers-core needs to be in the same directory as answers. The tests are failing in percy and circle because they do not have access to the core library.

The truncate function is the same as the truncate function in `ResultsFactory`. In future PRs, I expect to remove `ResultsFactory` completely

J=SLAP-878
TEST=manual

Build a sample site and perform universal searchers. Compare the results to the same site on the master branch and confirm the results are the same. Also manually inspect the resulting objects to verify correctness.