Use CoreAdapter and answers-core library for vertical searches (#1193)
Update the SDK to use the answers-core library as the network layer. Update the SearchDataTransformer and its test cases. Removed old "from" static methods from the models if they were unused as a result of this change.

TEST=manual,unit

Create a simple page with SearchBar and VerticalResults components. Log the output in the console and see the models are the expected shape. Also see the results appear properly in the VerticalResults DOM. Also ran the unit tests with npm run test and confirmed all tests pass.