globalstorage: 2.0 scaffold + tests (#1208)
Add the scaffolded global storage to the SDK (for storage 2.0). Add
tests for the new functions assuming new functionality. Note: The unit
tests are not expected to pass at this time as the global storage has
not been implemented yet. The linter is supposed to pass, though.

All tests from globalstorage are added except for the setAll and the
previous implementation of getAll (which was really a getAllWithPrefix).

J=SLAP-867
TEST=unit

* globalstorage: 2.0 scaffold + tests

Add the scaffolded global storage to the SDK (for storage 2.0). Add
tests for the new functions.

J=SLAP-867
TEST=unit

* Add tests for getUrlWithCurrentState

* Add newline bt jsdoc desc and param

* Forgot one

* Fix fn name, fix fn call in tests

* Test fixes after initial implementation