core-adapter: add autocomplete core adapter
Updates core-adapter.js to use the core library's
autocomplete functions.

Adds an `AutoCompleteResponseTransformer` class to
transform responses from the core library into a format
that the SDK understands and uses.

J=884
TEST=manual
Set up a test site to use my local version of the SDK
with the core adapter and saw that the autocomplete data
that is being generated via the core library and transformer
is equivalent to that being generated via just the SDK.