core-adapter: don't replace autocomplete result filter (#1214)
core-adapter: don't replace autocomplete result filter

Attempting to assign result.filter another value in
the autocompleteresponsetransformer results in an
error because result.filter is readonly.

This makes the change so that in the response transformer,
we pass data from the result into the constructor for
AutoCompleteResult in a new object that contains
transformed filters and the rest of the data from the result, instead of attempting to reassigned result.filter.

Also passes the sessionTrackingEnabled field to the
autocomplete request.

TEST=auto, manual
Ran automated tests and saw them pass.
On a local test site, verified that the readonly error
is gone and that the results have been transformed
corrected. Also using the same site, verified that
sessionTrackingEnabled is being passed as a request
paramter to the backend.