core-adapter: more tests for autocompleteresponsetransformer
Adds more test cases for autocompleteresponsetransformer.js.

Makes a change in autocompleteresponsetransformer.js to not
pass in `result.shortValue` to the AutoCompleteResult constructor
in `_transformAutoCompleteResult` since it is no longer part
of the response from the core library.

J=SLAP-941
TEST=auto
Ran tests with `npm run test` and saw that all of them passed.