pass queryTrigger: suggest when clicking spell check (#1230)
This ensures that we pass the queryTrigger: suggest param
after clicking a spell check link. Previously, this would
be overriden to query-parameter

This also fixes a bug where clicking a spell check link would
not reset the pagination from say page 2 to page 1. So if I were
to search for "varginia", go to page 2, then click the spell check,
prior to this I would be on page 2 instead of page 1 (though this
is probably exceptionally rare).

J=SLAP-979
TEST=manual, auto

test that clicking an autocomplete link will send queryTrigger: suggest

added acceptance test for spell check flow on a vertical page