Fix search bar buttons on very small screens (#1243)
Fix the search bar styling so that the buttons don't disappear on small screens

On screens with very small widths (<350px width), the search bar clear and search buttons shrink and eventually disappear as the screen width decreases. This PR prevents the buttons from shrinking and disappearing on very small screens.

J=SLAP-999
TEST=visual

Build a site and view it at various sizes between 300px and 400px wide and confirm that the search bar buttons no longer shrink or disappear. Test on a browswerstack iphone 8 with safare and on a windows laptop on IE11.