storage: use new setEntries for resetListener and post-search
This commit updates the current persistent storage resetListener
to also reseed the new storage in addition with the old storage.

To do this a new setEntries() method was added to the new storage,
which will set all entries given BEFORE calling any listeners, so
we don't need to worry about the order of storage listeners are called
anymore (see logic around StorageKeys.QUERY in  current
GlobalStorage's setAll()).
This was also added to core.search() and core.verticalSearch().

It also refactors the new storage listeners to be a map of
storage key to listener, so that setEntries() doesn't do a
double for loop when calling listeners.

J=SLAP-869
TEST=manual,auto

test that after searching for "virginia", then searching for "technology",
then hitting backspace, that ANSWERS.core.storage.get('query') returns
"virginia"

unit test that makes sure setEntries() will only call listeners